rename file to `specialized_ctrl`
