+++ qualtran/bloqs/chemistry/hubbard_model/qubitization/select_hubbard.py
@@ -191,7 +191,7 @@ class SelectHubbard(SelectOracle):
         return s
 
     def get_ctrl_system(self, ctrl_spec: 'CtrlSpec') -> Tuple['Bloq', 'AddControlledT']:
-        from qualtran.bloqs.mcmt.bloq_with_specialized_single_qubit_control import (
+        from qualtran.bloqs.mcmt.specialized_ctrl import (
             get_ctrl_system_for_bloq_with_specialized_single_qubit_control,
         )
