+++ qualtran/bloqs/chemistry/thc/select_bloq.py
@@ -315,7 +315,7 @@ class SelectTHC(SelectOracle):
         return out_soqs
 
     def get_ctrl_system(self, ctrl_spec: 'CtrlSpec') -> Tuple['Bloq', 'AddControlledT']:
-        from qualtran.bloqs.mcmt.bloq_with_specialized_single_qubit_control import (
+        from qualtran.bloqs.mcmt.specialized_ctrl import (
             get_ctrl_system_for_bloq_with_specialized_single_qubit_control,
         )
