+++ qualtran/bloqs/for_testing/random_select_and_prepare.py
@@ -150,7 +150,7 @@ class TestPauliSelectOracle(SelectOracle):  # type: ignore[misc]
             yield op
 
     def get_ctrl_system(self, ctrl_spec: 'CtrlSpec') -> Tuple['Bloq', 'AddControlledT']:
-        from qualtran.bloqs.mcmt.bloq_with_specialized_single_qubit_control import (
+        from qualtran.bloqs.mcmt.specialized_ctrl import (
             get_ctrl_system_for_bloq_with_specialized_single_qubit_control,
         )
