+++ qualtran/bloqs/mcmt/specialized_ctrl_test.py
@@ -42,7 +42,7 @@ from qualtran import (
     SoquetT,
 )
 from qualtran.bloqs.mcmt import And
-from qualtran.bloqs.mcmt.bloq_with_specialized_single_qubit_control import (
+from qualtran.bloqs.mcmt.specialized_ctrl import (
     get_ctrl_system_for_bloq_with_specialized_single_qubit_control,
 )
 from qualtran.resource_counting import CostKey, GateCounts, get_cost_value, QECGatesCost
