+++ qualtran/bloqs/reflections/reflection_using_prepare.py
@@ -195,7 +195,7 @@ class ReflectionUsingPrepare(GateWithRegisters):
         return self
 
     def get_ctrl_system(self, ctrl_spec: 'CtrlSpec') -> Tuple['Bloq', 'AddControlledT']:
-        from qualtran.bloqs.mcmt.bloq_with_specialized_single_qubit_control import (
+        from qualtran.bloqs.mcmt.specialized_ctrl import (
             get_ctrl_system_for_bloq_with_specialized_single_qubit_control,
         )
